Generative-AI/Generate-Image/requirements.txt
openai~=1.60.2
python-dotenv~=1.0.0
azure-identity>=1.15.0
azure-ai-projects